# Samat_Check/FBA/portal/fba/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "FBA POS API",
    "version": "1.0"
  },
  "host": "FBAappuv01.fortebank.com:8082",
  "paths": {
...
    "/receipt": {
      "get": {
        "description": "Получение списка чеков",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Receipt"
        ],
        "operationId": "ReceiptGET",
        "parameters": [
          {
            "type": "int",
            "description": "номер страницы пагинатора",
            "name": "page",
            "in": "query"
          },
          {
            "type": "int",
            "description": "количество записей на странице",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON массив чеков",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Receipt"
              }
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "description": "Изменение чека",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Receipt"
        ],
        "operationId": "ReceiptPUT",
        "parameters": [
          {
            "description": "JSON для изменения модели чека",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Receipt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON ответ на изменение чека",
            "schema": {
              "$ref": "#/definitions/Receipt"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Создание чека",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Receipt"
        ],
        "operationId": "ReceiptPOST",
        "parameters": [
          {
            "description": "JSON для создания чека",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Receipt"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON созданная модель чека",
            "schema": {
              "$ref": "#/definitions/Receipt"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/receipt/{id}": {
      "get": {
        "description": "Получение чека по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Receipt"
        ],
        "operationId": "ReceiptGETByID",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор чека",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель чека",
            "schema": {
              "$ref": "#/definitions/Receipt"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Удаление чека по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Receipt"
        ],
        "operationId": "ReceiptDELETE",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор чека",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/receipts/count": {
      "get": {
        "description": "Получение количества чеков",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Receipt"
        ],
        "operationId": "ReceiptCount",
        "responses": {
          "200": {
            "description": "JSON с количеством городов",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/tax": {
      "get": {
        "description": "Получение списка налогов",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tax"
        ],
        "operationId": "TaxGET",
        "parameters": [
          {
            "type": "int",
            "description": "номер страницы пагинатора",
            "name": "page",
            "in": "query"
          },
          {
            "type": "int",
            "description": "количество записей на странице",
            "name": "size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "JSON массив налогов",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Tax"
              }
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "description": "Изменение налога",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tax"
        ],
        "operationId": "TaxPUT",
        "parameters": [
          {
            "description": "JSON для изменения модели налога",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Tax"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON ответ на изменение налога",
            "schema": {
              "$ref": "#/definitions/Tax"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Создание налога",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tax"
        ],
        "operationId": "TaxPOST",
        "parameters": [
          {
            "description": "JSON для создания налога",
            "name": "bodyjson",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Tax"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JSON созданная модель налога",
            "schema": {
              "$ref": "#/definitions/Tax"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/tax/{id}": {
      "get": {
        "description": "Получение налога по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tax"
        ],
        "operationId": "TaxGETByID",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор налога",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "JSON модель налога",
            "schema": {
              "$ref": "#/definitions/Tax"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Удаление налога по ID",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tax"
        ],
        "operationId": "TaxDELETE",
        "parameters": [
          {
            "type": "string \u003cuuid4\u003e",
            "description": "идентификатор налога",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/taxs/count": {
      "get": {
        "description": "Получение количества налогов",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Tax"
        ],
        "operationId": "TaxCount",
        "responses": {
          "200": {
            "description": "JSON с количеством налогов",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          },
          "500": {
            "description": "JSON ответ об ошибке",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "City": {
      "description": "City справочник городов",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Company": {
      "description": "Company компании",
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "Bin": {
          "type": "string"
        },
        "Contact": {
          "type": "string"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Count": {
      "description": "modelRowsCount структура для количества записей под swagger",
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Count"
        }
      },
      "x-go-name": "modelRowsCount",
      "x-go-package": "fba/model"
    },
...
    "Error": {
      "description": "Error Ошибка",
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "fba/model"
    },
...
    "Point": {
      "description": "Point точки продаж",
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "City": {
          "$ref": "#/definitions/City"
        },
        "CityID": {
          "type": "string",
          "format": "uuid4"
        },
        "Company": {
          "$ref": "#/definitions/Company"
        },
        "CompanyID": {
          "type": "string",
          "format": "uuid4"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "ProductAction": {
      "description": "ProductAction Журнал движения товаров",
      "type": "object",
      "properties": {
        "ActionDate": {
          "type": "string",
          "format": "date-time"
        },
        "Amount": {
          "type": "number",
          "format": "double"
        },
        "Count": {
          "type": "integer",
          "format": "int64"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Point": {
          "$ref": "#/definitions/Point"
        },
        "PointID": {
          "type": "string",
          "format": "uuid4"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        },
        "WorkSession": {
          "$ref": "#/definitions/WorkSession"
        },
        "WorkSessionID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "Receipt": {
      "description": "Receipt чек",
      "type": "object",
      "properties": {
        "Checkdate": {
          "type": "string",
          "format": "date-time"
        },
        "Content": {
          "type": "string"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Number": {
          "type": "integer",
          "format": "int64"
        },
        "ProductAction": {
          "$ref": "#/definitions/ProductAction"
        },
        "ProductActionID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Role": {
      "description": "Role права пользователей(staff)",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "Staff": {
      "description": "Staff пользователь",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "ParentID": {
          "type": "string",
          "format": "uuid4"
        },
        "Role": {
          "$ref": "#/definitions/Role"
        },
        "RoleID": {
          "type": "string",
          "format": "uuid4"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
...
    "Tax": {
      "description": "Tax налоги",
      "type": "object",
      "properties": {
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Name": {
          "type": "string"
        },
        "ShortName": {
          "type": "string"
        },
        "ValidFrom": {
          "type": "string",
          "format": "date-time"
        },
        "Value": {
          "type": "number",
          "format": "float"
        }
      },
      "x-go-package": "fba/dbmodel"
    },
    "WorkSession": {
      "description": "WorkSession смена",
      "type": "object",
      "properties": {
        "Endamount": {
          "type": "number",
          "format": "double"
        },
        "Enddate": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "string",
          "format": "uuid4"
        },
        "Isactive": {
          "type": "boolean"
        },
        "Name": {
          "type": "string"
        },
        "Point": {
          "$ref": "#/definitions/Point"
        },
        "PointID": {
          "type": "string",
          "format": "uuid4"
        },
        "Staff": {
          "$ref": "#/definitions/Staff"
        },
        "StaffID": {
          "type": "string",
          "format": "uuid4"
        },
        "Startamount": {
          "type": "number",
          "format": "double"
        },
        "Startdate": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "fba/dbmodel"
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}